#64: Move UserRole classes to role package
