+++ src/main/java/org/codenergic/theskeleton/role/impl/RoleServiceImpl.java
@@ -16,10 +16,16 @@
 package org.codenergic.theskeleton.role.impl;
 
 import java.util.Objects;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 import org.codenergic.theskeleton.role.RoleEntity;
 import org.codenergic.theskeleton.role.RoleRepository;
 import org.codenergic.theskeleton.role.RoleService;
+import org.codenergic.theskeleton.role.UserRoleEntity;
+import org.codenergic.theskeleton.role.UserRoleRepository;
+import org.codenergic.theskeleton.user.UserEntity;
+import org.codenergic.theskeleton.user.UserRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -28,10 +34,22 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @Transactional(readOnly = true)
 public class RoleServiceImpl implements RoleService {
-	private RoleRepository roleRepository;
+	private final RoleRepository roleRepository;
+	private final UserRepository userRepository;
+	private final UserRoleRepository userRoleRepository;
 
-	public RoleServiceImpl(RoleRepository roleRepository) {
+	public RoleServiceImpl(RoleRepository roleRepository, UserRepository userRepository, UserRoleRepository userRoleRepository) {
 		this.roleRepository = roleRepository;
+		this.userRepository = userRepository;
+		this.userRoleRepository = userRoleRepository;
+	}
+
+	@Override
+	@Transactional
+	public UserEntity addRoleToUser(String username, String roleCode) {
+		UserEntity user = userRepository.findByUsername(username);
+		RoleEntity role = roleRepository.findByCode(roleCode);
+		return userRoleRepository.save(new UserRoleEntity(user, role)).getUser();
 	}
 
 	private void assertRoleNotNull(RoleEntity role) {
@@ -70,6 +88,21 @@ public class RoleServiceImpl implements RoleService {
 	@Override
 	public Page<RoleEntity> findRoles(String keywords, Pageable pageable) {
 		return roleRepository.findByCodeOrDescriptionStartsWith(keywords, pageable);
+	}
+
+	@Override
+	public Set<RoleEntity> findRolesByUserUsername(String username) {
+		return userRoleRepository.findByUserUsername(username).stream()
+			.map(UserRoleEntity::getRole)
+			.collect(Collectors.toSet());
+	}
+
+	@Override
+	@Transactional
+	public UserEntity removeRoleFromUser(String username, String roleCode) {
+		UserRoleEntity userRole = userRoleRepository.findByUserUsernameAndRoleCode(username, roleCode);
+		userRoleRepository.delete(userRole);
+		return userRepository.findByUsername(username);
 	}
 
 	@Override
