+++ src/main/java/org/codenergic/theskeleton/user/UserAdminService.java
@@ -16,29 +16,15 @@
 package org.codenergic.theskeleton.user;
 
 import java.util.Date;
-import java.util.Set;
 
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
 
-import org.codenergic.theskeleton.role.RoleEntity;
-import org.codenergic.theskeleton.privilege.RolePrivilegeRepository;
-import org.codenergic.theskeleton.role.RoleRepository;
-import org.codenergic.theskeleton.user.impl.UserServiceImpl;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.crypto.password.PasswordEncoder;
 
 public interface UserAdminService {
-	static UserAdminService newInstance(PasswordEncoder passwordEncoder, RoleRepository roleRepository, UserRepository userRepository,
-			UserRoleRepository userRoleRepository, RolePrivilegeRepository rolePrivilegeRepository) {
-		return new UserServiceImpl(passwordEncoder, roleRepository, userRepository, userRoleRepository, rolePrivilegeRepository);
-	}
-
-	@PreAuthorize("hasAuthority('user_assign_role')")
-	UserEntity addRoleToUser(@NotNull String username, @NotNull String roleCode);
-
 	@PreAuthorize("hasAuthority('user_delete')")
 	void deleteUser(@NotNull String username);
 
@@ -48,17 +34,11 @@ public interface UserAdminService {
 	@PreAuthorize("hasAuthority('user_update')")
 	UserEntity extendsUserExpiration(@NotNull String username, int amountInMinutes);
 
-	@PreAuthorize("hasAuthority('user_assign_role')")
-	Set<RoleEntity> findRolesByUserUsername(@NotNull String username);
-
 	@PreAuthorize("hasAuthority('user_read_all')")
 	Page<UserEntity> findUsersByUsernameStartingWith(@NotNull String username, Pageable pageable);
 
 	@PreAuthorize("hasAuthority('user_update')")
 	UserEntity lockOrUnlockUser(@NotNull String username, boolean unlocked);
-
-	@PreAuthorize("hasAuthority('user_assign_role')")
-	UserEntity removeRoleFromUser(@NotNull String username, @NotNull String roleCode);
 
 	@PreAuthorize("hasAuthority('user_write')")
 	UserEntity saveUser(@NotNull @Valid UserEntity userEntity);
