+++ src/main/java/org/codenergic/theskeleton/user/UserEntity.java
@@ -17,6 +17,7 @@ package org.codenergic.theskeleton.user;
 
 import org.codenergic.theskeleton.core.data.AbstractAuditingEntity;
 import org.codenergic.theskeleton.privilege.RolePrivilegeEntity;
+import org.codenergic.theskeleton.role.UserRoleEntity;
 import org.springframework.security.core.userdetails.UserDetails;
 
 import javax.persistence.*;
