+++ src/test/java/org/codenergic/theskeleton/role/RoleServiceTest.java
@@ -15,10 +15,14 @@
  */
 package org.codenergic.theskeleton.role;
 
-import java.util.Arrays;
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.Set;
 import java.util.UUID;
 
 import org.codenergic.theskeleton.role.impl.RoleServiceImpl;
+import org.codenergic.theskeleton.user.UserEntity;
+import org.codenergic.theskeleton.user.UserRepository;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
@@ -39,11 +43,44 @@ public class RoleServiceTest {
 	private RoleService roleService;
 	@Mock
 	private RoleRepository roleRepository;
+	@Mock
+	private UserRepository userRepository;
+	@Mock
+	private UserRoleRepository userRoleRepository;
 
 	@Before
 	public void init() {
 		MockitoAnnotations.initMocks(this);
-		this.roleService = new RoleServiceImpl(roleRepository);
+		this.roleService = new RoleServiceImpl(roleRepository, userRepository, userRoleRepository);
+	}
+
+	@Test
+	public void testAddRoleToUser() {
+		RoleEntity role = new RoleEntity()
+			.setId(UUID.randomUUID().toString())
+			.setCode("role");
+		UserEntity user = new UserEntity()
+			.setId(UUID.randomUUID().toString())
+			.setUsername("user");
+		UserRoleEntity result = new UserRoleEntity(user, role);
+		result.setId(UUID.randomUUID().toString());
+		when(roleRepository.findByCode("role")).thenReturn(role);
+		when(userRepository.findByUsername("user")).thenReturn(user);
+		when(userRoleRepository.save(any(UserRoleEntity.class))).thenReturn(result);
+		assertThat(roleService.addRoleToUser("user", "role")).isEqualTo(user);
+		verify(roleRepository).findByCode("role");
+		verify(userRepository).findByUsername("user");
+		verify(userRoleRepository).save(any(UserRoleEntity.class));
+	}
+
+	@Test
+	@SuppressWarnings("serial")
+	public void testDeleteRole() {
+		RoleEntity input = new RoleEntity() {{ setId("123"); }}.setCode("user");
+		when(roleRepository.findByCode("123")).thenReturn(input);
+		roleService.deleteRole("123");
+		verify(roleRepository).findByCode("123");
+		verify(roleRepository).delete(input);
 	}
 
 	@Test
@@ -74,10 +111,26 @@ public class RoleServiceTest {
 	@SuppressWarnings("serial")
 	public void testFindRoles() {
 		RoleEntity result = new RoleEntity() {{ setId("123"); }}.setCode("user");
-		Page<RoleEntity> page = new PageImpl<>(Arrays.asList(result));
+		Page<RoleEntity> page = new PageImpl<>(Collections.singletonList(result));
 		when(roleRepository.findAll(any(Pageable.class))).thenReturn(page);
 		assertThat(roleService.findRoles(new PageRequest(1, 10))).isEqualTo(page);
 		verify(roleRepository).findAll(any(Pageable.class));
+	}
+
+	@Test
+	public void testFindRolesByUserUsername() {
+		Set<UserRoleEntity> dbResult =
+			new HashSet<>(Collections.singletonList(new UserRoleEntity().setRole(new RoleEntity().setCode("role"))));
+		when(userRoleRepository.findByUserUsername("user")).thenReturn(dbResult);
+		Set<RoleEntity> result = roleService.findRolesByUserUsername("user");
+		assertThat(result.size()).isEqualTo(1);
+		assertThat(result.iterator().next()).isEqualTo(dbResult.iterator().next().getRole());
+		verify(userRoleRepository).findByUserUsername("user");
+	}
+
+	@Test
+	public void testRemoveRoleFromUser() {
+		roleService.removeRoleFromUser("", "");
 	}
 
 	@Test
@@ -103,15 +156,5 @@ public class RoleServiceTest {
 		assertThat(result.getId()).isEqualTo(input.getId());
 		assertThat(result.getCode()).isEqualTo(input.getCode());
 		verify(roleRepository).findByCode(eq("123"));
-	}
-
-	@Test
-	@SuppressWarnings("serial")
-	public void testDeleteRole() {
-		RoleEntity input = new RoleEntity() {{ setId("123"); }}.setCode("user");
-		when(roleRepository.findByCode("123")).thenReturn(input);
-		roleService.deleteRole("123");
-		verify(roleRepository).findByCode("123");
-		verify(roleRepository).delete(input);
 	}
 }
